WatchDog + XQuery cache integration
svn path=/trunk/eXist-1.0/; revision=330


Former-commit-id: 52484e6698892fa208a2971fbee946cf5dd289fa

+++ samples/org/exist/examples/xmlrpc/Store.java
@@ -38,7 +38,7 @@ public class Store {
 		Vector params = new Vector();
 		params.addElement(xml.toString());
 		params.addElement(args[1]);
-		params.addElement(new Integer(1));
+		params.addElement(new Integer(0));
 
 		// execute the call
 		Boolean result = (Boolean)xmlrpc.execute("parse", params);
